pyproject.toml
[tool.poetry]
name = "parse-companies"
version = "0.1.0"
description = ""
authors = ["Dimanskr <dimanskr2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
